+++ EVA00/ueva/uevafunctions.h
@@ -31,7 +31,7 @@ int masksOverlap(Mat &mask1, Mat &mask2);
 
 int detectKink(vector< Point_<int>> &contour, const int &convexSize);
 
-int detectNeck(vector< Point_<int>> &contour, int &kinkIndex, double &neck);
+int detectNeck(vector< Point_<int>> &contour, int &kinkIndex, float &neck);
 
 bool isCombinationPossible(vector<int> &combination, vector<UevaCtrl> &ctrls);
